feature_pipeline/requirements.txt
requests==2.31.0
pandas==2.0.3
hopsworks==4.1.5
python-dotenv==1.0.0
confluent-kafka>=2.3.0